requirements.txt
Adafruit-Blinka==5.4.0
adafruit-circuitpython-ads1x15==2.2.3
adafruit-circuitpython-busdevice==5.0.1
Adafruit-PlatformDetect==2.17.0
Adafruit-PureIO==1.1.5
asgiref==3.2.10
board==1.0
cassandra-driver==3.24.0
click==7.1.2
colorzero==1.1
Django==3.1.1
django-cassandra-engine==1.6.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
geomet==0.2.1.post1
gpiozero==1.5.1
pyftdi==0.51.2
pyserial==3.4
python-decouple==3.3
pytz==2020.1
pyusb==1.1.0
six==1.15.0
sqlparse==0.3.1
